chore: funding 이름 pick으로 변경 (#108)
