+++ src/app/wishpool/(viewer)/final/page.tsx
@@ -1,11 +1,11 @@
 'use client';
 import { useRouter } from 'next/navigation';
 
-import { data } from '@/app/funding/preview/data';
+import { data } from '@/app/pick/preview/data';
 import Button from '@/components/common/Button';
 import KakaoButton from '@/components/common/Button/KakaoButton';
 import Icon from '@/components/common/Icon';
-import GiftCard from '@/components/funding/list/GiftCard';
+import GiftCard from '@/components/pick/list/GiftCard';
 import { PATH } from '@/constants/common/path';
 
 const FinalPage = () => {
