+++ src/components/pick/select/SwipeGuide.tsx
