+++ src/constants/common/path.ts
@@ -21,12 +21,12 @@ export const PATH = {
   WISHPOOL_GIFTS: (id: number) => `/wishpool/${id}/gifts`,
   WISHPOOL_INVITE: (id: number) => `/wishpool/${id}/invite`,
 
-  FUNDING_INVITE: '/funding/invite',
-  FUNDING_INTRO: '/funding/intro',
-  FUNDING_SELECT: '/funding/select',
-  FUNDING_LIST: '/funding/list',
-  FUNDING_PREVIEW: '/funding/preview',
-  FUNDING_COMPLETE: '/funding/complete',
+  PICK_INVITE: '/pick/invite',
+  PICK_INTRO: '/pick/intro',
+  PICK_SELECT: '/pick/select',
+  PICK_LIST: '/pick/list',
+  PICK_PREVIEW: '/pick/preview',
+  PICK_COMPLETE: '/pick/complete',
 
   JOIN_INFO: `/wishpool/join/info`,
   JOIN_INTRO: `/wishpool/join/intro`,
